+++ docs/release-notes.md
@@ -1,6 +1,7 @@
 # Release notes
 ## 2.4
 * For players:
+  * SMAPI now fixes curly quotes in `config.json` if possible.
   * Fixed rare issues caused by assembly references being incorrectly loaded twice.
 
 * For the [log parser][]: